requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-generativeai==0.2.2
numpy==1.24.2
pandas==1.5.3
requests==2.28.2
yfinance==0.2.18
langchain==0.0.347